Split core.go file into logical subparts
Splitting the primary core.go file into more logical sub-sections which
will facilitate in easier code maintenance and readability in the
future.
